pyproject.toml
[project]
name = "direct-typer"
version = "0.1.0"
description = "macOS tool for typing text directly without clipboard"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyobjc-framework-Quartz>=10.0",
    "pynput>=1.7.6",
    "pyperclip>=1.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
